Split Writers.jl into multiple files
In order accommodate multiple output formats the Writers module should
be in multiple files. In order for the internal code for each format
not to conflict with other code, it should be wrapped in submodules.
